P2P: Fix cancelServiceDiscovery mis-cast
cancelServieDiscovery was mis-parsing/casting a long into an int, and
throwing a NumberFormatException.

Bug: 38245447
Test: CTS Verifier, updated unit tests

+++ service/java/com/android/server/wifi/p2p/SupplicantP2pIfaceHal.java
@@ -1405,9 +1405,9 @@ public class SupplicantP2pIfaceHal {
                 return false;
             }
 
-            int id = 0;
+            long id = 0;
             try {
-                id = Integer.parseInt(identifier);
+                id = Long.parseLong(identifier);
             } catch (NumberFormatException e) {
                 Log.e(TAG, "Service discovery identifier invalid: " + identifier, e);
                 return false;
